requirements.txt
asgiref==3.9.0
beautifulsoup4==4.13.4
Django==4.2.23
django-bootstrap-v5==1.0.11
django-bootstrap4==25.1
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
